% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spillovers.R
\name{advance_periods}
\alias{advance_periods}
\title{Add \code{n} quarters (or periods) to a \code{c(year, period)} date}
\usage{
advance_periods(yq, n, frequency = 4)
}
\description{
The inverse of \code{\link[=prev_period]{prev_period()}}, generalised to \code{n} steps.
}
\keyword{internal}
