% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_dbnomics.R
\name{dbnomics_cache_dir}
\alias{dbnomics_cache_dir}
\title{Directory used to cache DBnomics responses}
\usage{
dbnomics_cache_dir()
}
\description{
Directory used to cache DBnomics responses
}
\keyword{internal}
