% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_eamdqd.R
\name{eamdqd_bai_ng}
\alias{eamdqd_bai_ng}
\title{Select the number of factors via the Bai & Ng (2002) IC2 criterion}
\usage{
eamdqd_bai_ng(X, qmax)
}
\arguments{
\item{X}{Standardised \verb{T x N} numeric matrix, no missing values.}

\item{qmax}{Maximum number of factors to consider.}
}
\value{
Integer, the selected number of factors (\verb{>= 1}).
}
\description{
Select the number of factors via the Bai & Ng (2002) IC2 criterion
}
\keyword{internal}
