% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panel_build.R
\name{eamdqd_concept_codes}
\alias{eamdqd_concept_codes}
\title{Target variable set: project concept -> EA-MD/QD series code}
\usage{
eamdqd_concept_codes
}
\description{
Target variable set: project concept -> EA-MD/QD series code
}
\keyword{internal}
