% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panel_build.R
\name{eurostat_cache_dir}
\alias{eurostat_cache_dir}
\title{Directory used to cache Eurostat table downloads}
\usage{
eurostat_cache_dir()
}
\description{
Directory used to cache Eurostat table downloads
}
\keyword{internal}
