% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panel_build.R
\name{fred_concept_series}
\alias{fred_concept_series}
\title{Target variable set: project concept -> FRED series id}
\usage{
fred_concept_series
}
\description{
Target variable set: project concept -> FRED series id
}
\keyword{internal}
