% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panel_build.R
\name{get_custom_attrs}
\alias{get_custom_attrs}
\title{Recover the koma-specific attributes (series_type, method, ...) off a koma_ts}
\usage{
get_custom_attrs(x)
}
\description{
Recover the koma-specific attributes (series_type, method, ...) off a koma_ts
}
\keyword{internal}
