% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring.R
\name{log_score_gaussian}
\alias{log_score_gaussian}
\title{Log density of a Gaussian predictive distribution, evaluated at one point}
\usage{
log_score_gaussian(y, mu, sigma)
}
\arguments{
\item{y}{The realised value.}

\item{mu, sigma}{Mean and standard deviation of the predictive Gaussian.}
}
\value{
A scalar log score.
}
\description{
Log density of a Gaussian predictive distribution, evaluated at one point
}
