% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring.R
\name{origin_label}
\alias{origin_label}
\title{Label an origin for a cache filename or a progress message}
\usage{
origin_label(origin)
}
\description{
Label an origin for a cache filename or a progress message
}
\keyword{internal}
