% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{prev_period}
\alias{prev_period}
\title{Subtract one period from a \code{c(year, period)} date}
\usage{
prev_period(yq, frequency = 4)
}
\description{
Subtract one period from a \code{c(year, period)} date
}
\keyword{internal}
