% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stage1_models.R
\name{stage1_cache_dir}
\alias{stage1_cache_dir}
\title{Directory used to cache stage-1 fits}
\usage{
stage1_cache_dir()
}
\description{
Mirrors \code{\link[=eurostat_cache_dir]{eurostat_cache_dir()}}. Git-ignored via \verb{data/cache/*}.
}
\keyword{internal}
