% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panel_stage2d.R
\name{ts_ratio}
\alias{ts_ratio}
\title{Element-wise ratio of two \code{ts} over their common window}
\usage{
ts_ratio(numerator, denominator)
}
\description{
Element-wise ratio of two \code{ts} over their common window
}
\keyword{internal}
